Implement weather forecast

+++ UKWeather/src/com/haringeymobile/ukweather/database/CityTable.java
@@ -3,9 +3,9 @@ package com.haringeymobile.ukweather.database;
 import android.content.ContentValues;
 import android.database.sqlite.SQLiteDatabase;
 import android.provider.BaseColumns;
-import android.util.Log;
 
 import com.haringeymobile.ukweather.data.InitialCity;
+import com.haringeymobile.ukweather.utils.MiscMethods;
 
 public class CityTable implements BaseColumns {
 
@@ -15,19 +15,39 @@ public class CityTable implements BaseColumns {
 	public static final String TABLE_CITIES = "Cities";
 	public static final String COLUMN_CITY_ID = "Id";
 	public static final String COLUMN_NAME = "Name";
-	public static final String COLUMN_LAST_QUERY_DATE = "Date";
-	public static final String COLUMN_CACHED_JSON_CURRENT = "Current";
-	public static final String COLUMN_CACHED_JSON_FORECAST = "Forecast";
 
-	private static final String DATABASE_CREATE = "create table "
-			+ TABLE_CITIES + "(" + _ID + " integer primary key autoincrement, "
-			+ COLUMN_CITY_ID + " integer, " + COLUMN_NAME + " text not null, "
-			+ COLUMN_LAST_QUERY_DATE + " integer, "
+	public static final String COLUMN_LAST_QUERY_TIME_FOR_CURRENT_WEATHER = "TimeCurrent";
+	public static final String COLUMN_CACHED_JSON_CURRENT = "JsonCurrent";
+	public static final String COLUMN_LAST_QUERY_TIME_FOR_DAILY_WEATHER_FORECAST = "TimeDailyForecast";
+	public static final String COLUMN_CACHED_JSON_DAILY_FORECAST = "JsonDailyForecast";
+	public static final String COLUMN_LAST_QUERY_TIME_FOR_THREE_HOURLY_WEATHER_FORECAST = "TimeThreeHourlyForecast";
+	public static final String COLUMN_CACHED_JSON_THREE_HOURLY_FORECAST = "JsonThreeHourlyForecast";
+	public static final String COLUMN_LAST_OVERALL_QUERY_TIME = "LastQueryTime";
+
+	private static final String TABLE_TEMP = "tempTable";
+	private static final String COLUMN_LAST_QUERY_TIME_FOR_CURRENT_WEATHER_VERSION_1 = "Date";
+	private static final String COLUMN_CACHED_JSON_CURRENT_VERSION_1 = "Current";
+	// Not used in version 1:
+	// private static final String COLUMN_LAST_QUERY_TIME_FOR_WEATHER_FORECAST_VERSION_1 = "DateForecast";
+	// private static final String COLUMN_CACHED_JSON_FORECAST_VERSION_1 = "Forecast";
+
+	private static final String TABLE_CREATE = "CREATE TABLE "
+			+ TABLE_CITIES 
+			+ "(" 
+			+ _ID + " integer primary key autoincrement, "
+			+ COLUMN_CITY_ID + " integer, " 
+			+ COLUMN_NAME + " text not null, "
+			+ COLUMN_LAST_QUERY_TIME_FOR_CURRENT_WEATHER + " integer, "
 			+ COLUMN_CACHED_JSON_CURRENT + " text, "
-			+ COLUMN_CACHED_JSON_FORECAST + " text" + ");";
+			+ COLUMN_LAST_QUERY_TIME_FOR_DAILY_WEATHER_FORECAST + " integer, "
+			+ COLUMN_CACHED_JSON_DAILY_FORECAST + " text, "
+			+ COLUMN_LAST_QUERY_TIME_FOR_THREE_HOURLY_WEATHER_FORECAST + " integer, "
+			+ COLUMN_CACHED_JSON_THREE_HOURLY_FORECAST + " text, "
+			+ COLUMN_LAST_OVERALL_QUERY_TIME + " integer"
+			+ ");";
 
 	public static void onCreate(SQLiteDatabase database) {
-		database.execSQL(DATABASE_CREATE);
+		database.execSQL(TABLE_CREATE);
 		insertInitialData(database);
 	}
 
@@ -36,21 +56,71 @@ public class CityTable implements BaseColumns {
 			ContentValues newValues = new ContentValues();
 			newValues.put(COLUMN_CITY_ID, city.getOpenWeatherMapId());
 			newValues.put(COLUMN_NAME, city.getDisplayName());
-			newValues.put(COLUMN_LAST_QUERY_DATE, CITY_NEVER_QUERIED);
+			newValues.put(COLUMN_LAST_QUERY_TIME_FOR_CURRENT_WEATHER, CITY_NEVER_QUERIED);
 			newValues.putNull(COLUMN_CACHED_JSON_CURRENT);
-			newValues.putNull(COLUMN_CACHED_JSON_FORECAST);
+			putInitialDataForVersion2(newValues);
 			database.insert(TABLE_CITIES, null, newValues);
 		}
+	}
 
+	private static void putInitialDataForVersion2(ContentValues newValues) {
+		newValues.put(COLUMN_LAST_QUERY_TIME_FOR_DAILY_WEATHER_FORECAST, CITY_NEVER_QUERIED);
+		newValues.putNull(COLUMN_CACHED_JSON_DAILY_FORECAST);
+		newValues.put(COLUMN_LAST_QUERY_TIME_FOR_THREE_HOURLY_WEATHER_FORECAST, CITY_NEVER_QUERIED);
+		newValues.putNull(COLUMN_CACHED_JSON_THREE_HOURLY_FORECAST);
+		newValues.put(COLUMN_LAST_OVERALL_QUERY_TIME, CITY_NEVER_QUERIED);
 	}
 
 	public static void onUpgrade(SQLiteDatabase database, int oldVersion,
 			int newVersion) {
-		Log.w(CityTable.class.getName(), "Upgrading database from version "
-				+ oldVersion + " to " + newVersion
-				+ ", which will destroy all old data");
-		database.execSQL("DROP TABLE IF EXISTS " + TABLE_CITIES);
-		onCreate(database);
+		MiscMethods.log("Upgrading database from version " + oldVersion
+				+ " to version " + newVersion);
+		if (oldVersion == 1 && newVersion > 1) {
+			alterDatabaseVersion_1(database);
+		} else {
+			database.execSQL("DROP TABLE IF EXISTS " + TABLE_CITIES);
+			onCreate(database);
+		}
+	}
+
+	private static void alterDatabaseVersion_1(SQLiteDatabase database) {
+		database.beginTransaction();
+		try {
+			alterCityTable(database);
+			database.setTransactionSuccessful();
+		} catch (Exception e) {
+			throw new RuntimeException(
+					"Error upgrading database from version 1");
+		} finally {
+			database.endTransaction();
+		}
+	}
+
+	private static void alterCityTable(SQLiteDatabase database) {
+		String RENAME_ORIGINAL_TABLE = "ALTER TABLE " + TABLE_CITIES
+				+ " RENAME TO " + TABLE_TEMP;
+		String COPY_OLD_TABLE_TO_NEW_TABLE = "INSERT INTO " + TABLE_CITIES
+				+ "(" + COLUMN_CITY_ID + ", " + COLUMN_NAME + ", "
+				+ COLUMN_LAST_QUERY_TIME_FOR_CURRENT_WEATHER + ", "
+				+ COLUMN_CACHED_JSON_CURRENT + ") SELECT " + COLUMN_CITY_ID
+				+ ", " + COLUMN_NAME + ", "
+				+ COLUMN_LAST_QUERY_TIME_FOR_CURRENT_WEATHER_VERSION_1 + ", "
+				+ COLUMN_CACHED_JSON_CURRENT_VERSION_1 + " FROM " + TABLE_TEMP
+				+ ";";
+		
+		database.execSQL(RENAME_ORIGINAL_TABLE);
+		database.execSQL(TABLE_CREATE);
+		database.execSQL(COPY_OLD_TABLE_TO_NEW_TABLE);
+		database.execSQL("DROP TABLE " + TABLE_TEMP);
+		
+		insertInitialWeatherForecastValues(database);
+	}
+
+	private static void insertInitialWeatherForecastValues(
+			SQLiteDatabase database) {
+		ContentValues initialForecastValues = new ContentValues();
+		putInitialDataForVersion2(initialForecastValues);
+		database.insert(TABLE_CITIES, null, initialForecastValues);
 	}
 
 }
